+++ R/ui_helpers.R
@@ -80,11 +80,11 @@ app_brand_title <- function(version) {
 }
 
 app_stylesheet_link <- function(version) {
-  tags$link(rel = "stylesheet", type = "text/css", href = paste0("style.css?v=", version, "-analysis-mega-menu-20260620"))
+  tags$link(rel = "stylesheet", type = "text/css", href = paste0("style.css?v=", version, "-grouped-sample-effect-menus-20260620"))
 }
 
 app_script_link <- function(version) {
-  tags$script(src = paste0("easyflow.js?v=", version, "-analysis-mega-menu-20260620"))
+  tags$script(src = paste0("easyflow.js?v=", version, "-grouped-sample-effect-menus-20260620"))
 }
 
 app_head_tags <- function(version) {